Moving the timer out into it’s own separate file
